refactor: use interfeces, move them to separate file, add comments
